+++ src/senaite/core/upgrade/v02_06_000.py
@@ -1757,3 +1757,25 @@ def remove_category_title_metadata(tool):
     del_metadata(SETUP_CATALOG, "getCategoryTitle")
 
     logger.info("Removing getCategoryTitle metadata from catalogs [DONE]")
+
+
+def set_referenceable_behavior(tool):
+    """Assigns the referenceable behavior to setup folders so they are indexed
+    in the UID Catalog
+    """
+    behavior = "plone.app.referenceablebehavior.referenceable.IReferenceable"
+    logger.info("Assigning referenceable behavior to setup folders ...")
+    pt = api.get_tool("portal_types")
+    setup = api.get_senaite_setup()
+    to_fix = [setup] + list(setup.objectValues())
+    for obj in to_fix:
+        portal_type = api.get_portal_type(obj)
+        fti = pt.get(portal_type)
+        if behavior not in fti.behaviors:
+            logger.info("Adding IReferenceable behavior: %s" % portal_type)
+            behaviors = list(fti.behaviors) + [behavior]
+            fti.behaviors = tuple(behaviors)
+
+        obj.reindexObject()
+
+    logger.info("Assigning referenceable behavior to setup folders [DONE]")