+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #2571 Move InstrumentLocations to Senaite setup folder
+- #2578 Fix setup DX folders are not indexed in `uid_catalog`
 - #2576 Fix analyses and services are re-indexed after modifying category
 - #2577 Fix content types typo
 - #2567 Migrate AnalysisCategories to Dexterity